Add test all_ship_sunk?. cut and pasted play method to runner file.
